Added support for multiple ssl certificates
Added support to provide `haproxy.ssl_pem` as an array
of certs, to enable users to serve multiple https domains through
haproxy. If `haproxy.ssl_pem` is specified as a string, a single
cert is create, the same as before. If it is an array, each element
is treated as a separate cert, which gets loaded into haproxy.