Add option to accept new connections after SIGUSR1
Loadbalancers still try to open new connections after haproxy
shutdown is started with SIGUSR1. They only stop when they detect
the failing health check. Some loadbalancers do not retry opening
connections on other haproxy instances in this case but fail.

With the grace parameter it can be configured that the frontends
still are operational for some time, although the health endpoint fails.
This time should be defined in a way that the loadbalancer can detect
the unhealthy state.

[1] https://cbonte.github.io/haproxy-dconv/1.7/configuration.html#4.2-grace