New Mutual TLS setting compatible with CF Route Services
We have Haproxy terminating TLS and then Gorouter behind Haproxy. We want to use Mutual TLS
with Route Services. Right now this is broken because Gorouter makes an external request to
the route service, and Haproxy's `sanitize_set` option strips off the X-Forwarded-Client-Cert
header.

This commit adds a new setting `forwarded_client_cert: forward_only_if_route_service`.
In this mode Haproxy will preserve X-Forwarded-Client-Cert headers if there is also a
X-Cf-Proxy-Signature header provided.

The security of this is based on Gorouter verifying the contents of the X-Cf-Proxy-Signature
header. Gorouter checks the contents of that header and rejects the request if it did not
come from a CF Route Service (this works both for Gorouter->Route Service and
Route Service->Gorouter.)