New script to count the number of haproxy processes
We want easy access to the number of different haproxy processes,
but specifically the processes processing requests. We can have
several of these processes at a time after the `reload` script
has been called or when `nbproc` is >1.

In the following process tree there are 3 processes we're counting,
not the parent haproxy process. It's the 'middle' haproxy.

           └─tini(5706)─┬─haproxy(5747)─┬─haproxy(6313)─┬─{haproxy}(6314)
                        │               │               ├─{haproxy}(6315)
                        │               │               └─{haproxy}(6316)
                        │               ├─haproxy(6475)─┬─{haproxy}(6476)
                        │               │               ├─{haproxy}(6477)
                        │               │               └─{haproxy}(6478)
                        │               └─haproxy(6502)─┬─{haproxy}(6503)
                        │                               ├─{haproxy}(6504)
                        │                               └─{haproxy}(6505)

Co-authored-by: Peter Jones <peter.jones01@sap.com>