Treat all ACLs the same way in ERB templates
Rather than having a routed_backend_PREFIX rule then numbered rules
after that, just treat the path_beg rule as the first in a list and
combine it with the additional acls. This allows them all to be treated
the same way by building a hash of acl name: acl rule and using that in
templates. This shouldn't affect users as the job spec has not changed,
just the template within