Add cgroupsv1 check for Docker for Mac
Cancel local acceptance test attempt when there are changes in the repo
The script would eventually error out after a few minutes for the same reason.

Provide a script in the README.md that can be run by the operator to restart Docker for Mac with cgroupsv1 at their convenience.