fix(ci): replace spruce apt repository and install latest ginkgo
The apt repository `spruce` was in no longer exists. This commit removes
the logic to add it and installs `spruce` via `go install` instead.

Ginkgo has two versions: the installed one and the one defined in the
go.mod file. Instead of maintaining both this commit changes the version
installed in the container to latest.

Go binaries are now installed to `/usr/local/bin` to avoid messing with
the path by modifying `/root/.bashrc`.