fix(expect-proxy): provide alternative health endpoint for expect_proxy_cidrs
`expect_proxy_cidrs` is used for dynamically determining whether proxy protocol should be used or not.
On AWS the health check and regular traffic are forwarded with or without proxy protocol.

The idea of adding the `expect-proxy` directive CIDRs for AWS load balancer
IP addresses as a "marker" for traffic that should become Proxy Protocol, while
other traffic (including transparently proxied traffic) remains without
Proxy Protocol.

By adding a new health endpoint, when `expect_proxy_cidrs` is set, the
operator can select, which endpoint to use for checking the health of
HAProxy.