fix: properly check if TCP should be drained (#795)
When there is no HTTP health-check for a TCP frontend the drain logic
doesn't work as the frontend it tries to disable does not exist. This
commit makes the statement to disable the frontend during the draining
depend on whether the HTTP health-heck for that TCP frontend is enbaled
and if not excludes the drain logic for it.

Resolves: https://github.com/cloudfoundry/haproxy-boshrelease/issues/788